test: centralise helper setup_file_logging
