+++ rest-api/tests/integration_tests.rs
@@ -7,6 +7,7 @@ pub mod tests {
     use rest_api::server::run_server;
     use rest_api_auth::{HEADER_NONCE, HEADER_PUBLIC_KEY, HEADER_SIGNATURE, HEADER_TIMESTAMP};
     use rest_api_test_helpers::parse_log_lines;
+    use rest_api_test_helpers::setup_file_logging;
     use rest_api_test_helpers::wait_for_log_entry_with_request_id;
     use rest_api_test_helpers::{
         TestAuthHeaders, build_test_config, create_auth_headers, file_exists_in_repo,
@@ -17,11 +18,8 @@ pub mod tests {
     use rest_api_test_helpers::{file_exists_in_latest_commit, file_is_tracked_in_git};
     use serde_json::{Value, json};
     use std::fs;
-    use std::path::Path;
     use tempfile::TempDir;
     use tokio::time::Duration;
-    use tracing_appender::non_blocking::WorkerGuard;
-    use tracing_subscriber::{EnvFilter, Registry, fmt, layer::SubscriberExt};
 
     // Test case: Successfully add a file to the repository
     #[tokio::test]
@@ -373,26 +371,6 @@ pub mod tests {
     // ============================================================================
     // Structured Logging Tests
     // ============================================================================
-
-    fn setup_file_logging(
-        temp_dir: &Path,
-    ) -> Result<(
-        WorkerGuard,
-        std::path::PathBuf,
-        tracing::subscriber::DefaultGuard,
-    )> {
-        let log_path = temp_dir.join("test.log");
-        let appender = tracing_appender::rolling::never(temp_dir, "test.log");
-        let (non_blocking, guard) = tracing_appender::non_blocking(appender);
-
-        let subscriber = Registry::default()
-            .with(EnvFilter::new("info"))
-            .with(fmt::layer().json().with_writer(non_blocking));
-
-        let default_guard = tracing::subscriber::set_default(subscriber);
-
-        Ok((guard, log_path, default_guard))
-    }
 
     // Note: This test validates the basic setup of our logging (json format, timestamp,...).
     // It was generated by AI, and if it proves unnecessary can be removed.
