+++ rest_api_test_helpers/src/lib.rs
@@ -21,7 +21,8 @@ use tokio::{
     net::TcpStream,
     time::{Instant, timeout},
 };
-use tracing_subscriber::EnvFilter;
+use tracing_appender::non_blocking::WorkerGuard;
+use tracing_subscriber::{EnvFilter, Registry, fmt, layer::SubscriberExt};
 
 //
 // Helper function to initialize a git repository in a temporary directory
@@ -383,4 +384,24 @@ pub fn print_logs() {
         .with_env_filter(EnvFilter::from_default_env())
         .with_test_writer()
         .init();
+}
+
+pub fn setup_file_logging(
+    temp_dir: &Path,
+) -> Result<(
+    WorkerGuard,
+    std::path::PathBuf,
+    tracing::subscriber::DefaultGuard,
+)> {
+    let log_path = temp_dir.join("test.log");
+    let appender = tracing_appender::rolling::never(temp_dir, "test.log");
+    let (non_blocking, guard) = tracing_appender::non_blocking(appender);
+
+    let subscriber = Registry::default()
+        .with(EnvFilter::new("info"))
+        .with(fmt::layer().json().with_writer(non_blocking));
+
+    let default_guard = tracing::subscriber::set_default(subscriber);
+
+    Ok((guard, log_path, default_guard))
 }
